requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu
torchvision==0.12.0+cpu

orjson==3.8.0
python-multipart==0.0.5
uvicorn==0.18.3
gunicorn
fastapi
# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python-headless
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.41.0
protobuf<4.21.3
# Plotting ------------------------------------
pandas>=1.1.4
seaborn

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation